# api-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Demo classifieds marketplace API (C2C listings: items, categories, favorites, chat, listing promotion).\nNot affiliated with Avito, OLX, or any other classifieds platform.",
        "title": "Classifieds Marketplace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/items": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "List items catalog",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category UUID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by one or more category UUIDs, comma-separated",
                        "name": "category_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city UUID",
                        "name": "city_uuid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by GeoNames ID (alternative to city_uuid)",
                        "name": "geoname_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "new",
                            "used"
                        ],
                        "type": "string",
                        "description": "Filter by condition",
                        "name": "condition",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Min price in minor units",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max price in minor units",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by seller UUID",
                        "name": "seller_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by title",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated statuses (e.g. published,archived). Non-published only allowed for own seller profile.",
                        "name": "statuses",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_item.ItemListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Any authenticated user can create an item. Status is immediately 'published' (no moderation in MVP). Optional thumbnail_id sets the cover image from any uploaded item photo (independent of photo_ids).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create item",
                "parameters": [
                    {
                        "description": "Item data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_item.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api_item.ItemSingleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/api/v1/items/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item by ID or slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_item.ItemSingleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creator or admin can delete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Delete item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creator or admin can update. Admins are identified by ROLE_ADMIN in their JWT. The thumbnail_id is independent of photo_ids and can reference any uploaded item photo. To clear thumbnail, send thumbnail_id=null.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Update item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_item.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_item.ItemSingleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_dionisvl_classifieds_api-go_internal_api.PaginationResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_item.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.CityResponse": {
            "type": "object",
            "properties": {
                "geoname_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "names": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "population": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.CreateItemRequest": {
            "type": "object",
            "required": [
                "category_id",
                "title"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "city_uuid": {
                    "type": "string"
                },
                "condition": {
                    "type": "string",
                    "enum": [
                        "new",
                        "used"
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "geoname_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "photo_ids": {
                    "type": "array",
                    "maxItems": 10,
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "$ref": "#/definitions/internal_api_item.PriceRequest"
                },
                "tags": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "internal_api_item.ItemListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_item.ItemResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.PaginationResponse"
                }
            }
        },
        "internal_api_item.ItemResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/internal_api_item.CategoryResponse"
                },
                "category_id": {
                    "type": "string"
                },
                "city": {
                    "$ref": "#/definitions/internal_api_item.CityResponse"
                },
                "condition": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_favorited": {
                    "type": "boolean"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_item.PhotoResponse"
                    }
                },
                "price": {
                    "$ref": "#/definitions/internal_api_item.PriceResponse"
                },
                "seller": {
                    "$ref": "#/definitions/internal_api_item.SellerResponse"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "$ref": "#/definitions/internal_api_item.PhotoResponse"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.ItemSingleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_api_item.ItemResponse"
                }
            }
        },
        "internal_api_item.PhotoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.PriceRequest": {
            "type": "object",
            "required": [
                "currency"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.PriceResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.SellerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_api_item.UpdateItemRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "city_uuid": {
                    "type": "string"
                },
                "condition": {
                    "type": "string",
                    "enum": [
                        "new",
                        "used"
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "geoname_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "photo_ids": {
                    "type": "array",
                    "maxItems": 10,
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "$ref": "#/definitions/internal_api_item.PriceRequest"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "published",
                        "draft",
                        "archived",
                        "sold"
                    ]
                },
                "tags": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}